requirements.txt
# Core dependencies for Pigskin Auction Draft Tool
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
colorama>=0.4.6
pydantic-settings>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
